users can place orders, shopping cart is almost fully functional

+++ webshop/src/main/java/user/UserDao.java
@@ -13,7 +13,7 @@ public interface UserDao {
 	public AppUser selectUserByUserName(String username);
 	
 	
-	public int addUsers(AppUser user);
+	public Long addUsers(AppUser user);
 	
 	
 	public AppUser selectUserByUserId(Long id);
